Keep truncate_tests module rename
