+++ codex-rs/protocol/src/output_truncation.rs
@@ -184,4 +184,4 @@ pub fn approx_tokens_from_byte_count_i64(bytes: i64) -> i64 {
 }
 
 #[cfg(test)]
-mod tests;
+mod truncate_tests;
