+++ codex-rs/protocol/src/output_truncation/truncate_tests.rs
